Xong led và buzzer, dang fix bug buzzer

+++ robot-platform/main/src/Behavior/TurnRightBehavior.h
@@ -5,7 +5,7 @@
 
 class TurnRightBehavior : public Behavior {
 public:
-    TurnRightBehavior(int speed, uint16_t durationMs = 0);
+    TurnRightBehavior(int speed, uint32_t  durationMs = 0);
     virtual void init(BehaviorContext& context) override;
     virtual void start(BehaviorContext& context) override;
     virtual void update(BehaviorContext& context) override;
@@ -16,7 +16,7 @@ public:
 
 private:
     int speed;
-    uint16_t durationMs;
+    uint32_t durationMs;
     uint32_t startTime;
     bool isRunning;
 };
